requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data processing and analysis
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
PyYAML>=6.0.0

# Async support for concurrent API calls
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Logging and monitoring
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Data export
openpyxl>=3.1.0
